apps/terrarisk-agent/backend/pyproject.toml
[project]
name = "terrarisk-agent-backend"
version = "0.1.0"
description = "TerraRisk Agent backend service (FastAPI) for geospatial underwriting R&D."
authors = [{name = "Personal R&D"}]
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.111.0",
    "uvicorn[standard]>=0.29.0",
    "pydantic>=2.7.0",
    "pydantic-settings>=2.2.1",
    "httpx>=0.27.0",
    "google-cloud-bigquery>=3.20.0",
    "google-auth>=2.29.0",
    "geojson>=3.1.0",
    "shapely>=2.0.4",
    "pandas>=2.2.2",
    "numpy>=1.26.4",
    "jinja2>=3.1.3",
    "weasyprint>=62.3",
    "opentelemetry-sdk>=1.24.0",
    "opentelemetry-exporter-otlp>=1.24.0",
    "structlog>=24.1.0",
    "tenacity>=8.2.3"
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.4",
    "pytest-asyncio>=0.23.6",
    "ruff>=0.4.3",
    "mypy>=1.9.0",
    "types-requests"
]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.mypy]
ignore_missing_imports = true
strict_optional = true
